+++ src/Search/Search/SearchService/SetSearchServiceCommand.cs
@@ -16,6 +16,7 @@ using Microsoft.Azure.Commands.Management.Search.Models;
 using Microsoft.Azure.Commands.Management.Search.Properties;
 using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
 using Microsoft.Azure.Management.Internal.Resources.Utilities.Models;
+using Microsoft.Azure.Management.Search.Models;
 using System;
 using System.Management.Automation;
 
@@ -69,6 +70,16 @@ namespace Microsoft.Azure.Commands.Management.Search.SearchService
             HelpMessage = ReplicaCountHelpMessage)]
         public int? ReplicaCount { get; set; }
 
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "test")]
+        public PSPublicNetworkAccess? PublicNetworkAccess { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "test")]
+        public PSNetworkRuleSet NetworkRuleSet { get; set; }
+
         public override void ExecuteCmdlet()
         {
             if (ParameterSetName.Equals(InputObjectParameterSetName, StringComparison.InvariantCulture))
@@ -91,9 +102,24 @@ namespace Microsoft.Azure.Commands.Management.Search.SearchService
                     var service = SearchClient.Services.GetWithHttpMessagesAsync(ResourceGroupName, Name).Result.Body;
 
                     // UPDATE
-                    service.PartitionCount = PartitionCount;
-                    service.ReplicaCount = ReplicaCount;
-                    service = SearchClient.Services.UpdateWithHttpMessagesAsync(ResourceGroupName, Name, service).Result.Body;
+                    var updateService = new Microsoft.Azure.Management.Search.Models.SearchServiceUpdate()
+                    {
+                        // Keep existing properties
+                        HostingMode = service.HostingMode,
+                        Identity = service.Identity,
+                        Location = service.Location,
+                        //NetworkRuleSet = service.NetworkRuleSet,
+                        Sku = service.Sku,
+                        Tags = service.Tags,
+
+                        // Update the properties passed in
+                        NetworkRuleSet = (NetworkRuleSet)NetworkRuleSet,
+                        PublicNetworkAccess = (PublicNetworkAccess)PublicNetworkAccess,
+                        PartitionCount = PartitionCount,
+                        ReplicaCount = ReplicaCount
+                    };
+
+                    service = SearchClient.Services.UpdateWithHttpMessagesAsync(ResourceGroupName, Name, updateService).Result.Body;
 
                     // OUTPUT
                     WriteSearchService(service);